Separate routes for creating & searching user
